Add support for non persistent connections.
In an escenario where the wavefront proxy is in an ASG behind an ELB
where the proxy instances may change, is better to not have persistent
connections in order to ensure that the current one is always created
against a running instance. What is more, if you are using the
connection draining feature of the ELB's using non persistent
connections will ensure you that you can replace the instances in the
ASG without any data loss.

As the wavefront protocol doesn't expect to receive an ACK from the
remote server the client can't notice a dropped connection, a connection
that was not properly closed, and it can lead to send metrics to nowhere.